% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSJM-package.R
\docType{package}
\name{LSJM2-package}
\alias{LSJM2}
\alias{LSJM2-package}
\title{LSJM2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
